app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
python-dotenv==1.0.1
kubernetes==31.0.0
python-jenkins==1.8.2
requests==2.32.3
pydantic==2.9.2
httpx==0.27.2